changin format for files
